feat: add sharable module
