+++ .eslintrc
@@ -12,7 +12,11 @@
         "@typescript-eslint/explicit-module-boundary-types": 0,
         "@typescript-eslint/no-explicit-any": 0,
         "@typescript-eslint/no-unused-vars": [
-            "error"
+            "error",
+            {
+                "ignoreRestSiblings": false,
+                "varsIgnorePattern": "^_"
+            }
         ],
         "no-tabs": 0,
         "node/no-new-require": 0,
