+++ packages/shared/index.ts
@@ -1,0 +1,1 @@
+export { shared } from './src/index'
