+++ packages/shared/rollup.config.js
@@ -1,0 +1,6 @@
+import config from '../../build/builder'
+import { dependencies } from './package.json'
+
+export default config({
+    external: dependencies
+})
